Clean up and document the Rust code
